Refactor CMA-ES script for maintainability
Refactor `mtuq/stochastic_sampling/cmaes.py` by splitting it into multiple modules for better maintainability.

* **Add `mtuq/stochastic_sampling/cmaes_utils.py`**:
  - Add utility functions for CMA-ES, including `linear_transform`, `inverse_linear_transform`, `logarithmic_transform`, `in_bounds`, `array_in_bounds`, and `Repair` class.

* **Add `mtuq/stochastic_sampling/cmaes_initialization.py`**:
  - Add `CMAESInitialization` class to handle initialization tasks, including MPI communicator setup, logging, parameter initialization, and cache setup.

* **Add `mtuq/stochastic_sampling/cmaes_mutant_generation.py`**:
  - Add `CMAESMutantGeneration` class to handle mutant generation tasks, including drawing mutants, repairing out-of-bounds mutants, and scattering mutants across MPI processes.

* **Add `mtuq/stochastic_sampling/cmaes_fitness_evaluation.py`**:
  - Add `CMAESFitnessEvaluation` class to handle fitness evaluation tasks, including transforming mutants, generating sources, evaluating fitness, and handling different modes (db and greens).

* **Add `mtuq/stochastic_sampling/cmaes_plotting.py`**:
  - Add `CMAESPlotting` class to handle plotting tasks, including plotting mean waveforms and generating scatter plots of mutants and mean solutions.
